TCK progress
Buncha missing Int128 hooks, missing Int variants.  String split, indexof. Default args never used on some constructors. Missing space in "expr--28018L".  Boxed variants of methods do not unbox unboxable args.  MethodPart clz hoisted.  XValue missing imports.

+++ javatools_backend/src/main/java/org/xvm/xtc/XBase.java
@@ -39,7 +39,7 @@ public class XBase extends XType {
   // Because interning, if not "==" then never structurally equals and
   // because simple fully expanded types, never "isa" unless "==".
   @Override boolean _isa( XType xt ) {
-    if( this==XCons.LONG && xt==XCons.INT ) return true;
+    if( this==XCons.INT && xt==XCons.LONG ) return true;
     return S.eq( "String", _jtype ) && S.eq( "String", ((XBase) xt)._jtype ) && !xt._notNull;
   }
 }
